changed to only using the bcm library

+++ src/framework/raspberrypi/pal/timer-rpi.hpp
@@ -11,7 +11,8 @@
 #define TIMER_RPI_HPP_
 
 #include "../../../config/tle94112-conf.hpp"
-#include <wiringPi.h>
+// #include <wiringPi.h>
+#include <bcm2835.h>
 
 #if (TLE94112_FRAMEWORK == TLE94112_FRMWK_RPI)
 
@@ -30,13 +31,13 @@ class TimerRpi: virtual public Timer
 		Error_t init();
 		Error_t deinit();
 		Error_t start();
-		Error_t elapsed(uint32_t &elapsed); 
+		Error_t elapsed(uint64_t &elapsed); 
 		Error_t stop();
-		Error_t delayMilli(uint32_t timeout);
-		Error_t delayMicro(uint32_t timeout);
+		Error_t delayMilli(uint64_t timeout);
+		Error_t delayMicro(uint64_t timeout);
 
 	private:
-		uint32_t startTime; 
+		uint64_t startTime; 
 };
 /** @} */
